2021 11 09 성윤
이벤트 첨부파일 완성

+++ src/main/java/com/team4/joopging/event/vo/EventFileVO.java
@@ -10,5 +10,5 @@ public class EventFileVO {
     private String uploadPath;
     private String uuid;
     private boolean image;
-    private Long eventNum;
+    private int eventNum;
 }
